Add files via upload
Glove_ESP.ino - Reads positional values from a gyro-accelerometer and sends them over ESP_NOW. MAC address is currently configured to the esp in my (Brandon) system, so will need to be updated depending on the setup. Threshold values need to be configured properly.

main.cpp - Microros implementation that will receive arguments from Glove_ESP.ino, perform some logic, then publish to the moveit planner. Currently the values that are being published are set to increment depending on the information coming in from the glove. The initial pose messages are subject to change but must be within reasonable bounds for the arm or the planning will fail.

moveit_handler - Node used to begin planning in moveit. The tutorial demo for moveit is being used as our arm is modeled off of the panda arm used in the demo. In order for the node to run, the demo must be launched prior to the node. Once the plan is complete the node will publish the joint states of the arm.